+++ src/main/java/insane96mcp/enhancedai/mixin/ProjectileUtilMixin.java
@@ -1,0 +1,20 @@
+package insane96mcp.enhancedai.mixin;
+
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.monster.Drowned;
+import net.minecraft.world.entity.projectile.ProjectileUtil;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+
+@Mixin(ProjectileUtil.class)
+public abstract class ProjectileUtilMixin {
+    @WrapOperation(method = "getEntityHitResult(Lnet/minecraft/world/entity/Entity;Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/phys/Vec3;Lnet/minecraft/world/phys/AABB;Ljava/util/function/Predicate;D)Lnet/minecraft/world/phys/EntityHitResult;", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;canRiderInteract()Z"))
+    private static boolean enhancedai$canRiderInteract(Entity instance, Operation<Boolean> original) {
+        boolean result = original.call(instance);
+        if (instance instanceof Drowned)
+            return true;
+        return result;
+    }
+}