+++ src/main/java/insane96mcp/enhancedai/modules/mobs/AirSteal.java
@@ -1,0 +1,55 @@
+package insane96mcp.enhancedai.modules.mobs;
+
+import insane96mcp.enhancedai.EnhancedAI;
+import insane96mcp.enhancedai.data.EAIData;
+import insane96mcp.enhancedai.modules.Modules;
+import insane96mcp.insanelib.base.Feature;
+import insane96mcp.insanelib.base.LoadFeature;
+import insane96mcp.insanelib.base.Module;
+import insane96mcp.insanelib.base.config.Config;
+import net.minecraft.core.registries.Registries;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.Mob;
+import net.minecraftforge.event.entity.EntityJoinLevelEvent;
+import net.minecraftforge.event.entity.living.LivingHurtEvent;
+import net.minecraftforge.eventbus.api.SubscribeEvent;
+
+@LoadFeature(module = Modules.Ids.MOBS, description = "Only entity types in the enhancedai:mobs/air_stealer tag are affected by this feature.")
+public class AirSteal extends Feature {
+    public static final TagKey<EntityType<?>> AFFECTED_ENTITY_TYPES = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("mobs/air_stealer"));
+
+    @Config(min = 0, max = 128, description = "How many ticks of air are stolen from entities when attacked by the mob.")
+    public static Integer stolenTicks = 40;
+
+    public static EAIData<Integer> STOLEN_TICKS;
+
+    @Override
+    public void init(Module module, boolean enabledByDefault, boolean canBeDisabled) {
+        super.init(module, enabledByDefault, canBeDisabled);
+        STOLEN_TICKS = EAIData.ofInt(this.createDataKey("stolen_ticks"));
+    }
+
+    @SubscribeEvent
+    public void onEntityJoinLevel(EntityJoinLevelEvent event) {
+        if (!this.isEnabled()
+                || event.getLevel().isClientSide
+                || !(event.getEntity() instanceof Mob mob)
+                || !mob.getType().is(AFFECTED_ENTITY_TYPES))
+            return;
+
+        STOLEN_TICKS.applyIfAbsent(mob, stolenTicks);
+    }
+
+    @SubscribeEvent
+    public void onAttack(LivingHurtEvent event) {
+        if (!this.isEnabled()
+                || event.getEntity().level().isClientSide
+                || !(event.getSource().getEntity() instanceof LivingEntity attacker)
+                || STOLEN_TICKS.get(attacker) <= 0)
+            return;
+
+        event.getEntity().setAirSupply(event.getEntity().getAirSupply() - STOLEN_TICKS.get(attacker));
+    }
+}